pyproject.toml
[project]
name = "pydantic-ai-deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "pydantic-ai>=1.0.1",
    "pydantic-ai-litellm>=0.2.1",
    "python-dotenv>=1.0.0",
]
